create khmer/trimming.py, move trim_record there
