+++ scripts/filter-abund.py
@@ -60,6 +60,7 @@ from khmer.kfile import (check_input_files, check_space,
                          add_output_compression_type, get_file_writer)
 from khmer.khmer_logger import (configure_logging, log_info, log_error,
                                 log_warn)
+from khmer.trimming import (trim_record)
 
 DEFAULT_NORMALIZE_LIMIT = 20
 DEFAULT_CUTOFF = 2
@@ -114,26 +115,6 @@ def get_parser():
     return parser
 
 
-def trim_record(countgraph, record, variable_coverage, cutoff, normalize_to):
-    name = record.name
-    seq = record.sequence
-    seqN = seq.replace('N', 'A')
-
-    if variable_coverage:  # only trim when sequence has high enough C
-        med, _, _ = countgraph.get_median_count(seqN)
-        if med < normalize_to:
-            return name, seq
-
-    _, trim_at = countgraph.trim_on_abundance(seqN, cutoff)
-
-    if trim_at >= countgraph.ksize():
-        # be sure to not to change the 'N's in the trimmed sequence -
-        # so, return 'seq' and not 'seqN'.
-        return name, seq[:trim_at]
-
-    return None, None
-
-
 def main():
     args = sanitize_help(get_parser()).parse_args()
     if not args.quiet:
@@ -162,10 +143,6 @@ def main():
     if args.single_output_file:
         outfile = args.single_output_file.name
         outfp = get_file_writer(args.single_output_file, args.gzip, args.bzip)
-
-    def process_fn(record):
-        return trim_record(countgraph, record, args.variable_coverage,
-                           args.cutoff, args.normalize_to)
 
     # the filtering loop
     for infile in infiles:
